Moved pack subcommand to pack command
