+++ src/main/java/xyz/joestr/zonemenu/command/subcommand/SubCommandZoneCancel.java
@@ -1,4 +1,4 @@
-package xyz.joestr.zonemenu.subcommand;
+package xyz.joestr.zonemenu.command.subcommand;
 
 import org.bukkit.entity.Player;
